+++ CHANGES.md
@@ -6,6 +6,7 @@ quantities changelog
 
 - Made dimensionality printing nicer.
 - Added doctests and doc coverage tests.
+- Added unit conversion to fromString using "x => y"
 
 
 0.2 (4-6-2014)